+++ UBB--SE-2025-Workout-main/NeoIsisJob/Workout.Core/Repositories/UserRepository.cs
@@ -24,7 +24,7 @@ namespace Workout.Core.Repositories
             dbContext = context;
         }
 
-        public void JoinGroup(long userId, long groupId)
+        public void JoinGroup(int userId, long groupId)
         {
             try
             {
@@ -42,7 +42,7 @@ namespace Workout.Core.Repositories
             }
         }
 
-        public void ExitGroup(long userId, long groupId)
+        public void ExitGroup(int userId, long groupId)
         {
             try
             {
@@ -80,7 +80,7 @@ namespace Workout.Core.Repositories
         //    }
         //}
 
-        public void Follow(long userId, long whoToFollowId)
+        public void Follow(int userId, int whoToFollowId)
         {
             try
             {
@@ -90,7 +90,7 @@ namespace Workout.Core.Repositories
             catch { throw new Exception("Error following user"); }
 
         }
-        public void Unfollow(long userId, long whoToUnfollowId)
+        public void Unfollow(int userId, int whoToUnfollowId)
         {
             try
             {       
@@ -109,7 +109,7 @@ namespace Workout.Core.Repositories
         }
 
 
-        public UserModel GetById(long id)
+        public UserModel GetById(int id)
         {
             try
             {
@@ -131,7 +131,7 @@ namespace Workout.Core.Repositories
             }
         }
 
-        public List<UserModel> GetUserFollowers(long id)
+        public List<UserModel> GetUserFollowers(int id)
 
         {
             List<UserModel> userFollowers = new List<UserModel>();
@@ -150,7 +150,7 @@ namespace Workout.Core.Repositories
             return userFollowers;
         }
 
-        public List<UserModel> GetUserFollowing(long id)
+        public List<UserModel> GetUserFollowing(int id)
         {
             var userFollowers = dbContext.UserFollowers
                .Where(uf => uf.UserId == id);
